ML projects/face_detector/requirements.txt
matplotlib==3.1.1
Keras==2.3.1
pandas==0.24.2
opencv_python==4.1.2.30
numpy==1.17.3
tensorflow_gpu==2.0.0
Pillow==7.0.0
face_detector==0.4
tensorflow==2.1.0
